+++ metaengine/layout_calibration_bench_test.go
@@ -2,7 +2,7 @@ package metaengine_test
 
 import (
 	"context"
-	"encoding/json"
+	"encoding/json/v2"
 	"fmt"
 	"os"
 	"testing"